refactor: avoid creating lazy services on shutdown
